IMAGE SELECTION CODE
For the security.html, look at the CSS that has img1 - img10 and the div images with their hash-code specified. Also the on click function for each image is at the Bottom of the html. The hashing of the password is in the index3.js.  All of the images are provided.